SpelExpression implementation, removal of @UIEvent, removal of m-href
SpelExpression + EventHandlerRegistry makes for more stable expression parsing
To make it work within the context of an eventhandler/controller, it bases itself on the session's path
That works, but not with m-href, since the websocket needs to be 'reset' to connect with the right path. At which point it's probably just easier to do a normal href.
Additionally, @UIEvent becomes obsolete since SpelExpressions can pick up any method